texas-nlp/texas/requirements.txt
appnope
bleach
boto
boto3
botocore
bz2file
certifi
chardet
cycler
cymem
cytoolz
decorator
dill
docutils
elasticsearch
entrypoints
ftfy
gensim
html5lib
idna
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
jmespath
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
MarkupSafe
matplotlib
mistune
msgpack
msgpack-numpy
msgpack-python
murmurhash
nbconvert
nbformat
numpy
pandocfilters
parso
pathlib
pexpect
pickleshare
plac
preshed
prompt-toolkit
ptyprocess
Pygments
pyparsing
python-dateutil
pytz
pyzmq
qtconsole
regex
requests
s3transfer
scikit-learn
scipy
simplegeneric
six
smart-open
spacy
termcolor
terminado
testpath
thinc
toolz
tornado
tqdm
traitlets
ujson
urllib3
wcwidth
webencodings
widgetsnbextension
wrapt
